Add modp_b64 benchmark
modp_b64 is used in Chromium. It's twice as fast as fb64 at decode and
10% faster at encode, for 4× the decode footprint (4 kiB vs. 1 kiB for
fb64) and 6× the encode memory footprint (768 bytes vs. 128 bytes for
fb64).

modp_b64 source from
https://github.com/chromium/chromium/tree/master/third_party/modp_b64

It also over-allocates its output buffer by a few bytes (depending on
input length) and doesn't include base64url support. And the option of
padding has to be selected at compile-time. But it's faster.